// src/api/models/dto/CommentResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { DtoAuthorResponse } from "./AuthorResponse.ts";
import type { DtoMediaResponse } from "./MediaResponse.ts";
import type { DtoMentionResponse } from "./MentionResponse.ts";

export type DtoCommentResponse = {
    /**
     * @type object | undefined
    */
    author?: DtoAuthorResponse;
    /**
     * @type string | undefined
    */
    author_id?: string;
    /**
     * @type string | undefined
    */
    content?: string;
    /**
     * @type string | undefined
    */
    created_at?: string;
    /**
     * @type string | undefined
    */
    id?: string;
    /**
     * @type boolean | undefined
    */
    is_liked?: boolean;
    /**
     * @type integer | undefined
    */
    like_count?: number;
    /**
     * @type array | undefined
    */
    media?: DtoMediaResponse[];
    /**
     * @type array | undefined
    */
    mentions?: DtoMentionResponse[];
    /**
     * @type string | undefined
    */
    parent_id?: string;
    /**
     * @type string | undefined
    */
    post_id?: string;
    /**
     * @type array | undefined
    */
    replies?: DtoCommentResponse[];
    /**
     * @type integer | undefined
    */
    reply_count?: number;
};